Move files to Transactions folder
