--- e2e/selectors/Modals/TransactionProtectionModal.selectors.js
@@ -1,6 +1,0 @@
-import enContent from '../../../locales/languages/en.json';
-
-export const TransactionProtectionModalSelectorText = {
-  HEADER: enContent.whats_new.stx.header,
-  ENABLE_BUTTON: enContent.whats_new.stx.primary_button,
-};
