+++ e2e/selectors/Transactions/AssetWatcher.selectors.js
