+++ e2e/selectors/Transactions/TransactionDetailsModal.selectors.js
